pyproject.toml
[tool.poetry]
name = "chico-challanges"
version = "0.1.0"
description = "library api"
authors = ["chico"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mockito = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
